requirements.txt
langchain
GoogleGenerativeAI
requests
python-dotenv
